Add files via upload
As part of the IBM Data Science Certification Final Project, I applied my newly acquired skills to analyze a SpaceX dataset. My analysis focused on determining the likelihood of a first-stage landing, estimating the final cost of a launch, and evaluating whether the success of a launch depends on the location and proximity of the launch site.